+++ packages/terradart_google/lib/gke_backup.dart
@@ -9,19 +9,11 @@ export 'src/gke_backup/google_gke_backup_backup_channel.dart'
     show GoogleGkeBackupBackupChannel;
 export 'src/gke_backup/google_gke_backup_backup_plan.dart'
     show GoogleGkeBackupBackupPlan;
-export 'src/gke_backup/google_gke_backup_backup_plan_iam_binding.dart'
-    show GoogleGkeBackupBackupPlanIamBinding;
 export 'src/gke_backup/google_gke_backup_backup_plan_iam_member.dart'
     show GoogleGkeBackupBackupPlanIamMember;
-export 'src/gke_backup/google_gke_backup_backup_plan_iam_policy.dart'
-    show GoogleGkeBackupBackupPlanIamPolicy;
 export 'src/gke_backup/google_gke_backup_restore_channel.dart'
     show GoogleGkeBackupRestoreChannel;
 export 'src/gke_backup/google_gke_backup_restore_plan.dart'
     show GoogleGkeBackupRestorePlan;
-export 'src/gke_backup/google_gke_backup_restore_plan_iam_binding.dart'
-    show GoogleGkeBackupRestorePlanIamBinding;
 export 'src/gke_backup/google_gke_backup_restore_plan_iam_member.dart'
     show GoogleGkeBackupRestorePlanIamMember;
-export 'src/gke_backup/google_gke_backup_restore_plan_iam_policy.dart'
-    show GoogleGkeBackupRestorePlanIamPolicy;